requirements.txt
pandas==2.2.3
python-dotenv==1.1.0
python_docx==1.1.2
pywin32==310
tqdm==4.67.1
et-xmlfile=2.0.0
openpyxl=3.1.5
tabulate=0.9.0

